requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-autoslug==1.9.8
django-cities-light==3.8.1
django-cors-headers==3.7.0
django-extensions==3.1.1
django-import-export==2.5.0
django-jazzmin==2.4.4
django-oauth-toolkit==1.4.0
django-rest-framework-docs==0.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
jsonpickle==2.0.0
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
Pillow==8.1.0
pkg-resources==0.0.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dotenv==0.15.0
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
